requirements.txt
WTForms==3.0.1
SQLAlchemy==1.4.31
PyJWT==2.4.0
openpyxl==3.0.10
Flask==2.1.2
python-dotenv==0.20.0
convertapi==1.4.0
qrcode==7.3.1
PyPDF2==2.8.1
reportlab==3.6.11
aiosmtplib==1.1.6